+++ src/test/java/com/github/steingrd/bloominghollows/temperatures/TemperatureControllerTest.java
@@ -170,7 +170,7 @@ public class TemperatureControllerTest {
 			.andExpect(jsonPath("$", hasSize(2)));
 	}
 	
-	private static Temperature t(Brew brew, String date, int temp) {
+	private static Temperature t(Brew brew, String date, float temp) {
 		return new Temperature(brew, DateTime.parse(date), temp);
 	}
